Fix 64-bit compilation issues

+++ zbar/decoder.c
@@ -22,6 +22,7 @@
  *------------------------------------------------------------------------*/
 
 #include <config.h>
+#include <stdint.h>
 #include <stdlib.h>     /* malloc, calloc, free */
 #include <stdio.h>      /* snprintf */
 #include <string.h>     /* memset, strlen */
@@ -109,7 +110,7 @@ void zbar_decoder_destroy (zbar_decoder_t *dcode)
 
 void zbar_decoder_reset (zbar_decoder_t *dcode)
 {
-    memset(dcode, 0, (long)&dcode->buf_alloc - (long)dcode);
+    memset(dcode, 0, (intptr_t)&dcode->buf_alloc - (intptr_t)dcode);
 #ifdef ENABLE_EAN
     ean_reset(&dcode->ean);
 #endif
